refactor: created reusable chart controls
