+++ Expensier.WPF/Controls/Charts/ChartLegend.xaml
@@ -1,9 +1,9 @@
 ﻿<UserControl
-    x:Class="Expensier.WPF.Controls.ChartLegend"
+    x:Class="Expensier.WPF.Controls.Charts.ChartLegend"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:Expensier.WPF.Controls"
+    xmlns:local="clr-namespace:Expensier.WPF.Controls.Charts"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     HorizontalAlignment="Stretch"
     VerticalAlignment="Stretch"
