+++ Expensier.WPF/Controls/Charts/LineChart.xaml
@@ -1,0 +1,21 @@
+﻿<UserControl
+    x:Class="Expensier.WPF.Controls.Charts.LineChart"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:charts="clr-namespace:LiveChartsCore.SkiaSharpView.WPF;assembly=LiveChartsCore.SkiaSharpView.WPF"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="clr-namespace:Expensier.WPF.Controls.Charts"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    HorizontalAlignment="Stretch"
+    VerticalAlignment="Stretch"
+    mc:Ignorable="d">
+
+
+    <charts:CartesianChart
+        Height="280"
+        Series="{Binding Series}"
+        XAxes="{Binding XAxis}"
+        YAxes="{Binding YAxis}" />
+
+
+</UserControl>
