+++ Expensier.WPF/Controls/Charts/PieChart.xaml
@@ -1,0 +1,62 @@
+﻿<UserControl
+    x:Class="Expensier.WPF.Controls.Charts.PieChart"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:charts="clr-namespace:LiveChartsCore.SkiaSharpView.WPF;assembly=LiveChartsCore.SkiaSharpView.WPF"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="clr-namespace:Expensier.WPF.Controls.Charts"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    HorizontalAlignment="Stretch"
+    VerticalAlignment="Stretch"
+    mc:Ignorable="d">
+
+
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition Width="16" />
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
+
+        <!--#region Total-->
+        <Grid
+            Grid.Column="0"
+            HorizontalAlignment="Center"
+            VerticalAlignment="Center">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="8" />
+                <RowDefinition Height="Auto" />
+            </Grid.RowDefinitions>
+
+            <TextBlock
+                Grid.Row="0"
+                HorizontalAlignment="Center"
+                Style="{StaticResource Header6}"
+                Text="Total" />
+
+            <TextBlock
+                Grid.Row="2"
+                HorizontalAlignment="Center"
+                Style="{StaticResource Header4}"
+                Text="{Binding TotalExpenditure, StringFormat='C0'}" />
+        </Grid>
+        <!--#endregion-->
+
+        <!--#region Chart-->
+        <charts:PieChart
+            Grid.Column="0"
+            Width="280"
+            Height="280"
+            InitialRotation="-90"
+            Series="{Binding Series}" />
+        <!--#endregion-->
+
+        <!--#region Legend-->
+        <local:ChartLegend Grid.Column="2" />
+        <!--#endregion-->
+
+    </Grid>
+
+
+</UserControl>
